Significant changes in `Spiral.Main`. It seems that not using functions inside records is a good idea.
- Changed the StanState to accumulate nodes during initialization.
- Renamed the StanState to Context.
- Replaced the Rec and Layer records with functions. - Removed their ToArray fields.
- Remove the LayerWrapper class entirely.
- Context computational expression can now be used to chain layers.
- Added a generic UserState to the Context and remove Timestep.
- Added the with_userstate function.

All in all, I've removed more things than that I've added and I think the design is much better now. Parameterizing the layer creation function so that instead of randomly initializing the weight matrices they load from disk and other such features will not be a problem at all. It can be done by extending the Context to include commands.

Everything the Free monad can do, the Reader can do as well apart from resumable computations, but that can be done using the Delayed monad.

Apart from that the build is still broken. I'll move to the Combinators file next.