A non-update. I've made a resolution to go on an extended research trip away from this library for the sake of finally dealing with these shitty Cuda C string that are infesting it.
At first considered using translating F# quotation, which give me free parsing and type checking, to C code, but once I started it I realized very quickly that it would be difficult to figure out where to put the return statements from the AST.

I simply can't win like this.

So I am going to dedicate a large chunk of time, like a few months, to getting this particular aspect in order.

Current plan is to make a F# to LLVM compiler. The steps I have to take is to study various things:

1) FSCL.
2) LLVM.
3) GPU skeleton libraries.

And various other things. There is a massive amount of documentation for me to pore through and I can only do so much, so fast as a single person.

Until I get the pieces in order, I'll shed some burden by the wayside.