Agh, I did an enormous amount of research and gave it my best shot at learning Idris for the last ten days, but I gained nothing for it. I feel like I wasted the last 3 weeks.
While that is not true, I am not at all happy with what I attained. While Idris has potential, right now, doing anything non-trivial is too difficult in it.

So I will do an 180* turn and study reflection next. This is something that I've wanted to avoid since I am throwing away a lot of type safety there, but it is my only option now.

It did not seem like it 3 weeks ago, but I am starting to see the brighter side of it after bashing my head against the wall for so long.

As the first step, I'll take a look at Nemerle since it prides itself on its metaprogramming features and research .NET reflection in depth after that.

I'll do these modules in F# in the end.