Finally done with the map kernel using the new compiler. I ran into some issues described in this SO question: http://stackoverflow.com/questions/40423722/how-to-properly-pass-arguments-as-structs-to-nvrtc/
To get around the fact that F# always tuples the function arguments, I decided that for the kernel_main function to add a deforestation pass.

Another benefit of that the PTX will come out a lot better, but still quite horribly.

When I worked on the issue yesterday, I was horrified to see that with struct arguments, the compiler instead of just loading the individual field, will instead literally destrcture the struct down to individual bytes and then or and shift them using dozens of different registers just for the two field struct!

At this point, I can only cling to hope that the JIT will optimize this away.

F# and its quotations system really strikes me as not being the right tool for this job. I might give Idris another try now that I figured out how equality works in it.