I've finally completed the generic MapRedocolMap module using the compiler! To think that one month ago I simply wanted to implement layer normalization, but somehow ended up implementing a F# to Cuda compiler. How did that happen?
I am beat.

But to be honest, I need to proceed with this exercise in Idris as I am not satisfied with how it came out. While this thing undoubtedly works now, I am not going to even test it for performance, and in general, the structure of the program is not that good. It is the best I could do given the tools I have though.