Timed it, and the result is that the new compiler is 5x slower than the kernels I'd write by hand. Shit.
Basically, the JIT is doing absolutely nothing to make struct handling even remotely sensible. The generated PTX is complete garbage.

Ultimately, the problem is the AST that F# quotations give me. It is very unsuitable for translation to C. It seems that now, I either need to learn LLVM or Idris and probably both to get some measure of a good performance.

I'll plug away at Idris for a while longer.

I am really deeply enthralled by the idea of being able to take advantage of the compiler's type checking machinery in areas where it would be undoable in F# for example. It is a really powerful feature of dependently typed programming languages that does not exist anywhere else.