Disregard that last commit. It turns out the reason why it was so slow is because I set the grid size to 1 and block size to 32. I did that yesterday during debugging and it slipped my mind until now.
With that fixed the new compiler is only 20-30% slower because of that destructuring in the main function. Internalize it does optimize those tuples away.

Since this is the only overhead, there might be a point in further optimizing the main function.

It would be quite doable and quite typesafe.

But let me give tha a rest for now. I feel like another research adventure. I'll do this in a few days.