Did a little test to see if NVRTC can compile lambdas. No it can't. As expected.
Well, with that, now I have the green light to go with NVCC.

I will change the way I do programming. Looking forward, I can see that if I continue the way I used to do it, not even 10 of me would be enough to attain what I want.

Proofs are too heavy, and it is clear to me that I won't be able to attain the level of safety I desire with F#, so I'll change the way I look at it too. Instead of looking at it like a very safe static language, to push through to the next level, I will treat it as if I am programming in a dynamic one.

Any language can be used to cross the boundaries, and I F# would be well suited in the role traditionally taken by Python.

I'll take it easy for a while in order to refil my tank.