Gone in few steps further in making function to call nvcc from inside the library. Today when I looked at DeepNet it really surprised me to see that NVRTC does have C++11 support - for some reason I missed this when I went through the documentation in the past.
A change of focus to dynamism instead of safety and conciseness like up to now really helps clear the mind.

It is really too bad tuples are not supported natively. I tried including the Thrust library, but it would not work because it had a dependencies on stuff on the host side.

Since I do not have a way to get the compilation status through the command prompt, the way I implemented figuring when the compilation is finished in the function is by using the FileSystemWatcher class.

Not a move that I would even consider before because of its hackiness, but it should work quite nicely.

I'll add the new compilation function to the library proper tomorrow and use it to finish that generic RedoMapcolRedo.

It literally took me like 1.5 months just to do this simple thing, and now I will do it perfectly in just a three or four days.

I'll add some CUB functions to the mix too in order to speed up the reduction and replace that block reduce.

I wanted this for so long, and now I finally have it. Dynamism is great.

Type inference, safety, expresiveness...I had internalized three key aspects of programming those in the past one and a half year. Let me see if this new aspects will be enough to finally move on the offensive.