Thanks to a prompting by vitalygolub on SO, I decided to make an alternative version of the nvcc invoker that is better than what I had.
It has proper error detection now and is a lot less hacky that the previous solution. I really did make a mistake when I mentally weighted those two options and decided to go with the previous one first.

Ok, as far as calling nvcc is concerned, my needs are fully met now.

I'll just add the function to the main Spiral file and then, finish that fused column reduce kernel.

It is time I ended this thing and resume work on reinforcement learning. The clarion call of going to the next level with stronger static types simply too strong for me to resist.

I think now, I would be satisfied if F# gets typeclasses...and higher kinded types...and effect handlers. Yeah, those would be good enough for me.