It works quite well now. Also I am pleasantly surprised to see that auto can be the type of a lambda argument. I am also pleasantly surprised to find that the C++ compiler can infer the size of a local array as well.
Amazing.

This new code is not that much faster that the one produced by my F# Quotations compiler, but compared to the old kernels, it is in another league entirely.

I've pretty much gotten rid of that handmade block reduce with my own and replaced it with one that could work on any data type. I would be surprised to find it does not work on tuples.

Tomorrow I will do some more testing and then put it into the library proper. Then I will go back to completing that layer normalization thing.

Too long, it's been too long.