I've thought long and hard about this, and I think it is time I throw in the towel – on this iteration of the library that is. There is no doubt that it has some exquisitely elegant parts...but I am at my limit here.
It is impossible. I simply cannot express what I want with the way I am programming now.

I am bursting at the seams here.

I think it is time I properly embrace dynamism and take advantage of the "meta" of the "metalanguage" part from which F# inherits. By necessity, my code is going to have to start looking more like Lisp.

I've already had inkling that something was wrong when I could not figure out how to put in delayed evaluation into the program and had to resort to monads, but if I used the interpreter pattern from the start I would have had literally none of those problems.

Basically, the problem that I have here is that there is no way to implement Cuda map modules with variable number of arguments in the style I am programming now.

There is no way I will be able to interface the library with the world at large with the way it is designed now.

I thought I was making great progress, but in actuality I am in a very pernicious local minima.

In the next iteration, I think I'll me able to reuse most of the code from the current version in the evaluator.

The next iteration will definitely be a lot uglier and will have more type checking pushed to runtime, but in the end, I will not have the limits with it that I have now. This is not the best I can do.

Not even close. Let me do it once more, from the beginning.