I did some work today on the new intermediate language. About half of the new codegen is finished, and I'll do a little more work today after having a break.
I made a mistake with how I handle types here.

Really, having it in this form and starting from the codegen part really does unchain my mind.

It took me a while to to figure out, but the way to deal with grouped arrays either 1d or 2d is quite simple - make a type for them.

This is not something I could have done chained to the quotations system

This will work. This way of programming is the true path. I will not run into any more blocks while applying this style. I'll finish Spiral and then use it when I get back to reinforcement learning.

It is waiting for me.