Completed the generic map module.
This was an issue that plagued me for a long time and I am glad that I finally have Cuda under wraps now.

The function for it is generic in the number of arguments and accepts lambdas as well.

This is joy. At this point, I am starting to think that my code is no longer smart, but starting to approach genius.

Yeah, this is what I was trying to grasp. This kind of programming will form the backbone of my approach to real world problems from here on out.

Ever since the last Simulacrum update, I've tried to make my code safer and safer while trying to make it more expressive at the same time to no avail. When I realized I have a problem, I studied Haskell and dependetly typed programming, but what I really needed was to introduce some dynamism into the mix.

This is it. This is the point of maximum expressiveness and safety with ML-like functional languages.

Now that I did this, I need to think a little what the next step is. I am thinking of doing the dxM types now...