"Ouch, the bright ligh.
But seriously, nononono. What the hell am I doing here?

I had a good long think, and I've realized that I am doing everything wrong.
type DM =
    | Df of P: Lazy<float32> ref * A: float32 ref
    | D1M of n: int ref * diff: AutoDiffType ref * is_dead : DeadFlagType ref // flag to skip backprop
    | D2M of cr: (int * int) ref * diff: AutoDiffType ref * is_dead : DeadFlagType ref
    | D3M of ncr: (int * int * int) ref * diff: AutoDiffType ref * is_dead : DeadFlagType ref
    | D4M of nchw: (int * int * int * int) ref * diff: AutoDiffType ref * is_dead : DeadFlagType ref

It occured to me when I wrote brilliant things such as the above, but it just not where exactly did I go wrong.

Agh.

The way to get polymorphism with ADTs is different than when doing standard programming.

Shit.

And those object pools and workspaces, those should all be inside the evaluator. I need to push everything above, I need to push everything to the compiler level!

Right now I am letting the implementation and what I did previously dictate what the library should be.

Throw it away. Throw it all away and start from the beginning. There is a higher level than this."

Well, the above is an excerpt from the journal. I did a brilliant job with the module and the produced complete garbage with the library. I need to dump all of this and start afresh. I'll figure out how to reuse what I wrote in the previous library after that.