A non-update. In accordance with the effort to develop this new style, I think I will learn Lisp for a bit and see if I can learn any new techniques to use with F# discriminated unions.
Of course, I already studied Lisp before for a bit, but I never dove into it before, but this new style is pretty Lispy in substance, so it behooves me do it.

Hopefully I won't go away for like two weeks or more like when I was studying Haskell, Idris and Coq. I'll try to apply what I have learned.

Paul Graham's book On Lisp is of special interest to me since in the preword it promises to show the things that are possible in Lisp, but not in other languages.

What I did with those Cuda modules in just like 3 days is something I could not do for months before. And yesterday I realized that I really am way too attached to the old structure of the library.

During the night I was wondering whether I even need those DM types.

I'll do it properly this time - and if I finally manage to succeed, the way in front of me will be paved clean.