A non-update. I had enough of Lisp. The On Lisp book is interesting, and studying it properly rather than just skimming it would be quite an undertaking, but I do not think I'll be using Lisp.
But the essential lesson of Lisp is something that was lost on me before. If I want code that is both highly efficient and generic, the best thing would be to generate it in a DSL.

This is what Spiral should be.

Also I thought of generating F# code in Lisp, but I should really be generating F# code in F#.

I'll let this urge that is driving me forward consume me and maybe at the end, I will get something.

If I want to make my own language, the best bet would be to evolve it from here. That Cuda module was an omen of things to come. Rather than make a library, I need to evolve a set of primitives with which to interface with the world at large.