Ok, Racket seems like a very attractive choice right now. It would be far easier to write Cuda bindings in that language than adding new language features to F# instead.
I have't fully decided whether I want to leave F# greener pastures, but I want to take some time to try doing it in Typed Racket. I'll continue development in this repo as usual.

In the end, the only way I can relieve the pain of being stuck is by moving to higher places. I originally fell in love with F# because of its superiority to other languages and because mastering its functional language features gave such boosts to my programming skills, but if Racket can offer me more, then I will not hesitate to leave it behind and maybe make make a binding in order to fulfill my obligation to this library.

Only one way to find out.