I've come back to F# after studying Racket.
I intended to study macros, but instead what I did was read the [Programming Languages: Application and Interpretation](https://cs.brown.edu/~sk/Publications/Books/ProgLangs/2007-04-26/) book instead.

To be honest, I learned quite a bit. Before, I did not know the CPS conversion, now how the HM algorithm worked. And the thing on adding lazy semantics to the language might come useful in the Spiral library itself. It should definitely have parallel execution by default and laziness would help that.

It had some laziness in the last iteration in fact.

Even though Racket code is used within, it is an a static ML flavored dialect of it and most is done with pattern matching.

As I have not internalized the above subjects, I might do the exercises in F# at some point.

But yeah, although I am crushed that writting dynamic interpreters is the most powerful style, I will accept the curse of Lisp and move on.

Since it is tedious to redo the same fucking library for the fifth time in six months, it will take a me a while to build up the motivation to do it.

So I'll do it slowly as I said I would. Today I started off, but fixing a few bugs in the Cuda compiler part and significantly improved upon the syntax. I've also replaced Array1D and Array2D with a generic arrays whose dimensions are determined by the size of the arguments passed to it during declaration.

Tomorrow, I'll start work on the library core.