...It is not that this piece of code was difficult to write. It is figuring out what I wanted to write in the first place - the structure of the code that was difficult.
Done with the first part of the new object pool along with the new, more generic DM type.

Once can get a lot of genericity by simply using branches inside the getDM function.

The next would be to deal with the workspace. Unlike before, I think I will put it inside the object pool this time.