Did some work today, I'll do yet more later.
I defined the basic AST and am mostly trying to build a model in my head before I start coding in earnest.

Some of the things from the previous library like the deadness check have been removed, ids have been added to assist memoization and now I am thinking of how to handle branches.

Here is a cut from my journal reflecting my thinking on the current state of the library:

"Well, since I am first building up a tree, I no longer need to check for disconnected nodes – they simply won't be executed. I might need to build up a structure to check if all are fully connected.

But the other aspect of the deadness check really gave me pause. I had to think for a while exactly what I wanted to do with it.

Originally, I did not add the deadness check to look for disconnected nodes, but to act as an optimization so the unselected action layers do not get propagated. I never really got that far, but that is why I invented and put that particular function into the library. I got much more use for it as an assertion.

But looking at it right now, this selection of the the highest action layer...that is kind of like a case statement in a regular programming language, right?

Time and time again, I see the same abstractions manifest, just in different forms.

And I am thinking of whether the general form of my backprop is correct. I am using a stack now, but what if I wanted to implement those syntetic gradients ala Deepmind?

My hunch is that there is a way of generalizing the backwards step, I just do not know what it is.

I'll save figuring that out when I get to it."

I'll continue after lunch.