Preliminary commit. I split the print_arguments into get_method_arguments and print_arguments. Now using the first function, I'll create some type checking for arguments passed into the cuda modules.
Yesterday, I studied a bit TypeScript and about intersection types, but it was a dead end - my view that the writting dynamic DSLs in a static language like F# is the way to go has only been reinforced.

I want to spend the time until the end of the year working on the Spiral library. After that it is time for the update on the Simulacrum blog.

These last 5.5 months turned out to be a mess. My life really isn't long enough that I can continue to program in this manner.

What was the point on working on this library for so long is I am just going to end up using TF or CNTK? Saying this is for practice might be fine at first, but I need to stop practicing and start doing.