FIxed the type errors and a bug in the map_launcher function. That function has been completely redone to be significantly more generic compared to the last iteration of the library.
What I did today besides that is create the kernel_launcher function which performs type checking on the cuda arguments.

I'll do the same function for the other modules tomorrow. I am really tired right now, usually I would be able to go for another hour, but I will call it a day here as far as programming is concerned.

It is time for some Long War. I will take that trap harvester.