Save Point.
I made a mistake in optimizing the number of arguments in the backwards map function. It is more important that the functions have an uniform interface in order to keep the number of arguments to a minimum.

The backwards step should be a mirror of the forward step.

If I want to optimize out the unnecessary arguments, I should use a separate mechanism for that rather than do it manually.

The reason why this mirror form is necessary is because without it, I will have to create a seperate interface for each function. The activation_f that I made is already enough of a problem on its own.

Well, at any rate, the map functions having to take a few unnecessary arguments...I doubt that will really hurt performance in any noticeable way.

So I can just leave it out of mind.