Done with the first fifth of the evaluator.
Yeah, this is it - the hierarchy I created is perfect. All the pieces are neatly falling into place. It is amazing.

The map_backwards module that I made is sheer genius and doing that allowed me to reuse map_launcher even for the hadamarad multiplication.

Since my focus is on recurrent nets, I'll ommit the convolutional stuff. The code for that in the previous iteration of the library and I can add it whenever I need.

I'll ommit those parts for now and race towards getting everything ready for the cost functions.

After that I can bring back the Mnist tests and finally start work on recurrent nets.

I can't wait.