Did some preliminary work for the day by rederiving map_module and map_module_backwards functions from the mapcoef_module functions.
Also since the time is a approaching where I will need to type the evaluator to return multiple values, I put some time into understading how to emulate GADTs using standard DUs and the crucial bit of insight came to me!

http://stackoverflow.com/questions/40062180/how-can-this-phantom-type-example-possibly-be-valid/

I asked this two months ago, and only now it occured to me that an extrator function is necessary inside the DU to connect the evaluator to the intended structure. I was looking it with the wrong mindset back then.

Now I am just embarassed that I did not think of this myself.

With that out of the way, what is next on the agenda?