I fixed all the type errors and have successfuly added the tracking of dimensions to the library. With this, the new library is on par with the old one in that respect, except better.
I've also replaced generic_flatten_cudagroup with a functional version.

And as it turns out, I do not think it is possible to make generic_activations with DUs in the way I envisioned, so I ended up relying on closures. I was so focused in doing it that particular way and tired, that I did not think of this solution. A night's sleep had some benefits.

At any rate, today I mostly slacked - I do not think I put it 2h of work into this despite it being 3:15pm. Which is strange as it was not my intention to slack. I'll try to do at least 5h of work today.

The next thing to put in is Sum and the rest of the operations necessary for the cost function and then I will be able to finally write tests for the library.