It is like i am in hell. It is not even that I am unfocused, but just that nothing makes sense anymore to me.
It seems like I will have to redesign the generic_launcher. It seems that whatever abstractions I put in place so that map and map_reduce_map fall under one umbrella are failure.

In times like these, what is necessary is to put the things I've done into a separate file, get rid of the non-working pieces and start from a clean slate.

I can't deal with this.