Done polishing the cuda modules. Now everything flows much better together.
On review I just noticed that seqmatmult function lacks a  `Nodes.Add(id,c)` operation which is a bug.

I think that unifying the interface for the evaluator sounds like a swell idea right now. I'll do that tomorrow. Looking at it even further I can spot some other bugs in the rest of the function. The evaluator functions are full of roaches no matter how I look at it.

I need to remove the reduncancy from the code to get rid of them.

Today was a dream day. I was in the zone the whole time, and managed to resolve a problem I did not think was possible.