Done converting the evaluator to the tagless style. Originally I wanted to have each and every node memoized, but while doing this conversion I realized that it is completely unnecessary. So it all came out as regular code.
Instead of doing it as I intended before, what I will do is move memoization to the layer level. And then I will create a mini-language for expressing recurrence.

Doing it like that could also open the door modyfing the backward step as well, like with Deepmind like synthetic gradient methods.

But deciding to apply the tagless style is really a radical departure from my reason for redoing this library in the first place.

I was intenting to make the new library as a dynamic interpreter, but it is going to end up as a static compiler at this rate.

That is absolutely brilliant. Everything is falling into place for me - that new long sought for style is something that is coming into my reach.

So with that I feel like doing an another research adventure.

Though unlike before, this time it will be in this library.

I am going to try redoing the Cuda compiler with just functinos and no discriminated unions, in the tagless style. I am simply curious how it will turn out.

Too, too curious. In terms of functionality the thing I have now is perfectly fine, but I want to go forward and see if I can attain something new.

This will likely take me a few days.

Why would I do this? Why indeed.