Somehow I did manage to get in a bit of extra work.
As it turns out, I did not need the polymophism provided by the tagless style at all, but turning the outermost functions into lambdas instead of having them be directly evaluated as in the previous library will open up a new vista of opportunities for the design of the future parts.

Also, although it seem trivial, removing the ids from this level is also a good decision.

Somehow I was pretty stupid in terms of thought process today, but I think I made some pretty good decisions.

I'll stick to the decision to go tagless style even though it is just a lot of boilerplate at the moment. It may pay off in the future.

Since it is 7:30pm here, I think I'll call it a day here.