I refactored out the interpreters, fixed the bug where sigmoid was calling tanh and changed the names a little.
Yeah, this should be fine.

If I want to modify the library it won't be hard just by adding another level of hierarchy at this point right here.

On the last iteration I had the mistake of mixing closures and values, but now I have this extra necessary layer that will make it easy to change interpreters.

I'll have to regactor back to inline function, but the point where I will need to make a Futhark backend or something like that is far into the future.

I feel like apologizing to Athas, but at this point just doing what I set out to do in the first place takes precedence over everything else.