I am going through the old library right now, trying to figure out how to put the old into the new. Now that I have the cost functions, I think it is finally time to start working on tests.
I do not really like this part as no doubt a painful debugging session is ahead of me, due to not being attentive enough and missing an argument in some function somewhere.

It does not seem like I will be able to reuse much of the old stuff directly. As much effort as I put in into those combinators, I really have a superior system in place now and will have to adapt the old ideas afresh.

Agh, let me bring in the Mnist again and test the feedforward parts. I'll think of what to do next after that.