I added a few lines and with this the discriminated union evaluator is on par with the functional one.
Today, I slept the bost in a long time. And when I got up instead of working on the library, I spent it reading manga and learning Lisp for a bit.

Regarding Lisp, they make a big deal about the macro system, and indeed, you definitely can express more in Lisp than you can in F#.

But the point of having a type system is that it is a brain jack into what the compiler thinks. It allows you to offload thinking to the compiler.

With dynamic languages you still have to do type inference, except in your head.

Scala has implicits and Haskell has typeclasses, but the thing those features have in common is that they are kind of like type driven macros.

While in Lisp, macros are syntax driven, in highly advanced static languages, they are type driven. In the end, macros are just a type of polymorphism.

There is a point to be made that type inference in fact does help.

While studying Clojure (way back) I was surprised to find that Lisps do not have partial application by default.

But it makes sense.

Doing the functional composition of the kind one would do in F#, would be difficult in Lisp without the help from Intellisense.

That having said, I am really bored with the dynamic interpreter. I'll try to put in a few hours each day until I get over the hump.

I would not mind a language that had an extensible type system. It does not feel that great to be held at bay by the language's limitations. If I could add GADTs myself, I would do it.

I think for the future of Lisp, that would be an interesting direction to take it in.

How unmotivating. Let me try focusing at least for a little while, tomorrow I'll take some time to start writting up the update.