Done removing that useless total_size lambda that I stupidly added last night.
Now I can start work on the next step. I have no more doubts, I completely understand how to compose functions and how to emulate pattern matching using them.

It can be done similarly as with parser combinators. The trick is to define a separate mini-DSL for pattern matching on that structure and pass along the resulting function from that via a continuation.

With that the road is set.

I'll be busy with other things this week so I am not sure how much I'll be able to do, but I am going to get the library into usable form in the next few weeks.