Let me call it a day here.
Tomorrow's agenda:
- Clean up the Cuda segment.
- Push weights into the enviroment on layer creation.
- Make the runner deal with the backwards step.
- Make the train function.
- Test it on the nets Mnist to see if the new library works.

I did quite a bit on thinking on the above. I also have a now idea on how to convert the library to use CPS which would allow it to pattern match on layers with just functions. I'll defer that part until I actually need it. Right now it is only an idea, and I do not need it.