This thing of giving the layers runners is turning out to be more complicated than I thought. To be honest it is 7:25pm here and I do not have any energy to do it right now.
I got a head start today, but when I get that, usually I take that time to do sideprojects. Those range from potentially useful, like when I finally figured out the Y Combinator today and reading some of the ML papers that were recently posted on the ML sub, to useless surfing of social sites or blogs. I usually split it 50/50.

December was different from the previous months for me as I actually took up the long lost hobby of playing games after calling it a day. I am in the endgame in Long War.

I think in the future, Xcom: Enemy Within's mod Long War will become known as a classic.

I took the time to write the update for the Simulacrum the day before yesterday and tomorrow I'll proofread it and publish it.

Happy new year.