Finally put the post on https://gamesoftranscendi.wordpress.com/. I found a lot of typos while reading it after it was published.
That took me a while. And I intend to fuflil my vow this time. At the start of 2016 I set out to do poker, but in the end this library is still a bottleneck to me.

Let me get on with the show.

Before I do anything else, let me say that those squared_error_template and cross_entropy_template functions might be the greatest pieces of code I've ever written.

The end result is more code than had I packed it all into one function, but they provide so much clarity that it is incredible.

I want to explore the template pattern. I already stumpled upon it while doing the Cuda modules, but I screwed up here.

Looking this code, it is the complete opposite of those cuda functions. It is incredibly convoluted, and I have no idea what it is suppoed to do.

The template functions are supposed to be specifications, and the standard functions the implentation.

Just how long have did it have to take me to figure this out.

It was hard because of the extra code and the seemingly needless parameters.

I need to try it again.

Let me work on the Cuda modules for a bit. I already spent way too much on that post.

Let me Github trail serve as the proof of my efforts.