Done with the map_module and mapcoef_module functions.
This is it.

This is the exact partitioning of concerns that I could not reach before. The reason for that is that required a look at from a different perspective.

To be honest, back when I first found out about the Y Combinator there was no sense or logic or it no matter how long I thought about it, but now I figured it out so easily.

Ok, with that I feel like calling it a day.

I am actually pretty tired. To think I'd have to do more work on the Cuda compiler part, but after this I'll be done with it for sure.

Tomorrow I'll do the backwards parts and the rest of the modules and then I'll redo the functions.