My amazement at The Template Pattern never seems to cease. I started to reason about the Cuda modules that I've written from its perspective and I've realized that I've been writing them completely wrong. Absolutely fucking wrong.
It seems that the library is due for the sixth rewrite.

The great difference is that with the Template Pattern is that I am optimizing the specification and then the implementation while with everything else, I am doing both at the same time which is much harder.

Of course, I did not completely suck before, but I always had the sense that something was missing. And this is it. I lacked this principle. And if I am right adopting these new coding habits will allow me to take much longer and more accurate strides through the search space of programs.

To start with today, I am going to try redoing the modules with the template pattern to test the robustness of my new principles.