Uff, done with the first part of the rewrite. It is an interesting experiment and quite doable.
The Template Pattern can do anything, but to be honest, I need a break here. I am exhausted and can't write another line.

I need to recharge my energy.

After that I'll need to rerwite cuda_map_all to be more like cuda_map_module_compile.

Once I do that, I'll be able to write Cuda code the same way I would standard F# code. No need to put objects in lists and deal with all those indentation errors.

I'll be able to use let statements and variables directly.

It is a miracle. And array variables I'll be able to access by applying the accessors to them.

Quite amazing. Not only will the new compiler be purely functional, it will also be much more safer.

And I'll add signature hooks, so I'll be able to make statically typed calls into Cuda modules!

And by the looks of it, the code will be smaller too.

With this technique, no doubt I will be able to get to the level where I will be able to use RL to affect the real world. It has been a long time in coming.