I bashed the Haskell monads in my 2016 review, but they are what I need here.
Now that I think about it, most of what I got from the template pattern can also be gotten using the reader monad...well not really, but using the reader monad in tandem with it really looks like a good idea.

The template pattern is really an architectural optimization. It changed my way how I approach code, but there is no point in gorgin on it.

Now that I found my holy grail, I can time my time drinking from it lightly.

Starting from tomorrow, as a new year's resolution, I'll stop channeling insanity so much as I do not need it anymore.

Ultimately, if something needs to be done, and it cannot be done sanely, it will have to be done insanely, but I should be fine here.

I won't have to redo the library again and I won't have to deal with all this architectural nonsense.