Ok, I was not slacking today, but getting up 3 hours later than usual cost me quite a bit.
Did yesterday's session really tire me out that much?

At any rate I've figured it out.

The situation I am having now is similar to when I made GUIs in F#. The lack of dynamic binding is what did me in. I had to flatten everything.

And here the problem is different.

The local scoping is what is the problem.

Scala's implicits, typeclasses, effect handlers, what all these language features have in common is that they alter the perception of scope and in effect make it `implicit`.

The parameters in the template pattern should in fact be implicit and not gotten from the top level.

But if my claim that lexical scoping was a mistake, then to move to implicit scoping one would need a language with impredicative polymorphism.

Agh, that feature is something even Haskell has trouble with as it is so difficult. It is not even on the horizon, and like fish in water, I do not think anybody except me right now thinks implicit scoping is the problem.

Ack…

I'll do my best with what I have now. Maybe I should try Racket seriously at some point to see if it offers anything substantial.

Damn, it let me just do some proper programming. I'll roll back the changes and went to the earlier thing.

I'll separate statements and expressions, and unlike before I won't deal with the indentation right away, but at the end of compilation.

I'll do a trick, I'll define a type to especially help me with this.

Let me give it a shot. It does not look I'll be able to do much today, but I'll do my best.